+++ config/db.js
@@ -1,0 +1,9 @@
+module.exports = {
+  db: 'horizon',
+  servers: [
+    {
+      host: '127.0.0.1',
+      port: 28015
+    }
+  ]
+};