+++ source/server/app.js
@@ -4,10 +4,15 @@ import horizon from '@horizon/server';
 import devProps from '../../config/webpack/devProps';
 import config from '../../config/page';
 
+import './jobs';
+
 const app = express();
 
 app.use('/static', express.static(path.join(process.cwd(), '.build')));
 
+/**
+ * @TODO move the html out of the server dir
+ */
 const bundle = process.env.NODE_ENV === 'production'
 ? '/static/client.bundle.js'
 : `http://127.0.0.1:${devProps.webpackPort}/static/client.bundle.js`;
@@ -17,6 +22,7 @@ app.use('/', (req, res) => {
     <html>
       <head>
         <title>${config.title}</title>
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.6.1/css/font-awesome.min.css">
       </head>
       <body>
         <div id='root'></div>
@@ -37,7 +43,7 @@ const run = () => {
     console.log(`Express listening at http://localhost:${port}`); // eslint-disable-line
   });
 
-  // if we want to start secure, add key and cert props
+  // @TODO make this configurable
   const horizonServer = horizon(httpServer, {
     auto_create_table: true,
     auto_create_index: true,