+++ source/server/db.js
@@ -1,0 +1,6 @@
+import rethinkdbdash from 'rethinkdbdash';
+import config from '../../config/db';
+
+const r = rethinkdbdash(config);
+
+export default r;